+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,6 +1,7 @@
 Version X.X.X
 =============
  
+ * 2014-01-07 Security fix. Html templates did not escape apostrophes properly ('). (ra)
  * 2013-12-22 Added Maven 3.1.0 as requirement for ninja-maven-plugin (ra)
  * 2013-12-19 ObjectMapperProvider now a singleton (it is threadsafe) (ra)